Streamlit_app/requirements.txt
streamlit>=1.2.0
st-annotated-text==2.0.0
markdown>=3.3.4
farm-haystack==1.0.0
pandas==1.3.5
gdown==4.2.0